# gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.10.9-SNAPSHOT"
servers:
  - url: http://localhost:8083/portal/environments/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
  /apis:
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/apiContextPathQueryParam'
        - $ref: '#/components/parameters/apiLabelQueryParam'
        - $ref: '#/components/parameters/apiVersionQueryParam'
        - $ref: '#/components/parameters/apiNameQueryParam'
        - $ref: '#/components/parameters/apiCategoryQueryParam'
        - $ref: '#/components/parameters/apiFilterQueryParam'
        - $ref: '#/components/parameters/apiFilterExcludedQueryParam'
        - $ref: '#/components/parameters/apiPromotedQueryParam'
      summary: List APIs
      description: |
        List public APIs for anonymous requests.
        List all the APIs the current user is allowed to category for authenticated requests.

        The list can be filtered according to query parameters.

        By default the list is sorted by alphabetic order.
        If a **cat** query param is sent, a specific sorting can be applied. Please refer to **cat** description for detail.
      operationId: getApis
      security: []
      responses:
        200:
          description: List accessible APIs for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApisResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/_search:
    post:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/apiQueryParam'
      summary: Search APIs with a query
      description: |
        Same as `/apis` but with a search query as a parameter. Search for API using the search engine.
        Supports pagination.
      operationId: searchApis
      security: []
      responses:
        200:
          description: Search accessible APIs for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApisResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/{apiId}:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/apiIncludeParam'
      summary: Get the API definition
      description:  |
        Get the detail of an API.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getApiByApiId
      security: []
      responses:
        200:
          description: API definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /apis/{apiId}/picture:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      summary: Get the API's picture
      description: |
        Get the API's picture.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getPictureByApiId
      security: []
      responses:
        200:
          $ref: '#/components/responses/PictureSuccess'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /groups:
    get:
      tags:
        - Group
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      summary: List groups
      description: |
        List all groups

        User must have the MANAGEMENT_GROUP[READ] permission.
      operationId: getGroups
      responses:
        200:
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /categories:
    get:
      tags:
        - Portal
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      summary: Get a Category list
      description: |
        Get all categories of the platform.
      operationId: getCategories
      security: []
      responses:
        200:
          description: List of Categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  parameters:
...
    apiIdParam:
      name: apiId
      in: path
      required: true
      description: Id of an API.
      schema:
        type: string
...
    pageNumberParam:
      name: page
      in: query
      required: false
      description: The page number for pagination.
      schema:
        type: integer
        default: 1
    pageSizeParam:
      name: size
      in: query
      required: false
      description: |
        The number of items per page for pagination.
        If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
        If the size is -1, the response contains all datas.
      schema:
        type: integer
        default: 10

    #Apis filtering
    apiContextPathQueryParam:
      name: context-path
      in: query
      required: false
      description: The context-path of an API.
      schema:
        type: string
    apiLabelQueryParam:
      name: label
      in: query
      required: false
      description: One of the labels of an API.
      schema:
        type: string
    apiVersionQueryParam:
      name: version
      in: query
      required: false
      description: The version of an API.
      schema:
        type: string
    apiNameQueryParam:
      name: name
      in: query
      required: false
      description: The name of an API.
      schema:
        type: string
    apiCategoryQueryParam:
      name: category
      in: query
      required: false
      description: One of the category linked with an API.
      schema:
        type: string
    apiFilterQueryParam:
      name: filter
      in: query
      required: false
      description: |
        Name of a FILTER. A filter is used to pre-select some APIs. It can be combined with others query params.
          - FEATURED : top APIs. Sorted by specific order.
          - MINE : APIs subscribed by the current user. Sorted by alphabetical order.
          - STARRED: Rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
          - TRENDINGS: Subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
      schema:
        $ref: '#/components/schemas/FilterApiQuery'
    apiFilterExcludedQueryParam:
      name: -filter
      in: query
      required: false
      description: |
        Name of a FILTER to exclude. Excluded filter has priority over filter. A filter is used to pre-select some APIs. It can be combined with others query params.
          - FEATURED : not in top APIs. Sorted by specific order.
          - MINE : APIs not subscribed by the current user. Sorted by alphabetical order.
          - STARRED: not rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
          - TRENDINGS: not subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
      schema:
        $ref: '#/components/schemas/FilterApiQuery'
    apiPromotedQueryParam:
      name: promoted
      in: query
      required: false
      description: |
        Allow to filter on promoted APIs or not.
          - If absent, no effect on the result
          - If *false*, removes the promoted API of the list
          - If *true*, returns only the promoted API of the list

        A promoted API is either the highlighted API of a category (if it exists) or the first API in the list.
      schema:
        type: boolean
    apiQueryParam:
      name: q
      in: query
      required: true
      description: query string to be used in the search engine
      schema:
        type: string
    apiIncludeParam:
      name: include
      in: query
      required: false
      description: Comma-separated list of related objects to include in the response.
      schema:
        type: array
        items:
          type: string
          enum:
            - pages
            - plans
...
  schemas:
...
    ErrorResponse:
      properties:
        errors:
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
    ApisResponse:
      properties:
        data:
          description: List of API.
          type: array
          items:
            $ref: '#/components/schemas/Api'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    GroupsResponse:
      properties:
        data:
          description: List of groups.
          type: array
          items:
            $ref: '#/components/schemas/Group'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    CategoriesResponse:
      properties:
        data:
          description: List of categories.
          type: array
          items:
            $ref: '#/components/schemas/Category'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    Api:
      description: |
        Describes an API.
      required:
        - id
        - name
        - version
        - description
        - owner
      properties:
        id:
          description: Unique identifier of an API.
          type: string
        name:
          description: Name of the API.
          type: string
        version:
          description: Version of the API.
          type: string
        description:
          description: Description of the API.
          type: string
        draft:
          type: boolean
          description: Whether or not the API is in draft.
        public:
          type: boolean
          description: Whether or not the API is public.
        running:
          type: boolean
          description: Whether or not the API is running.
        entrypoints:
          description: List of all the avaible endpoints to call the API.
          type: array
          items:
            type: string
        labels:
          description: List of labels linked to this API.
          type: array
          items:
            type: string
        owner:
          $ref: '#/components/schemas/User'
        created_at:
          description: create date and time.
          type: string
          format: date-time
        updated_at:
          description: Last update date and time.
          type: string
          format: date-time
        categories:
          description: List of categories this API belongs to.
          type: array
          items:
            type: string
        rating_summary:
          $ref: '#/components/schemas/RatingSummary'
        _links:
          $ref: '#/components/schemas/ApiLinks'

        ###################
        # may be included #
        ###################
        pages:
          description: |
            Only returned with */apis/{apiId}*.
            Need *include* query param to contain 'pages'.

            The documentation pages of this API. Same as */apis/{apiId}/pages*.
          type: array
          items:
            $ref: '#/components/schemas/Page'
        plans:
          description: |
            Only returned with */apis/{apiId}*.
            Need *include* query param to contain 'plans'.

            The plans of this API. Same as */apis/{apiId}/plans*.
          type: array
          items:
            $ref: '#/components/schemas/Plan'
...
    Group:
      required:
        - id
        - name
      properties:
        id:
          description: Unique identifier of a group.
          type: string
        name:
          description: Name of the group.
          type: string
    Page:
      required:
        - id
        - name
        - type
        - order
      properties:
        id:
          description: Unique identifier of a page.
          type: string
        name:
          description: Name of the page.
          type: string
        type:
          description: Type of documentation.
          type: string
          enum:
            - ASCIIDOC
            - SWAGGER
            - MARKDOWN
            - FOLDER
            - ROOT
            - LINK
        order:
          description: Order of the documentation page in its folder.
          type: integer
        parent:
          description: Parent page. MAY be null.
          type: string
        updated_at:
          description: Last update date and time.
          type: string
          format: date-time
        configuration:
          $ref: '#/components/schemas/PageConfiguration'
        media:
          description: list of media hash, attached to this page
          type: array
          items:
            $ref: '#/components/schemas/PageMedia'
        metadata:
          description: Array of metadata about the page. This array is filled when the page has been fetched from a distant source (GitHub, GitLab, etc...).
          type: array
          items:
            $ref: '#/components/schemas/Metadata'
        _links:
          $ref: '#/components/schemas/PageLinks'

        ###################
        # may be included #
        ###################
        content:
          description: |
            Only returned with */apis/{apiId}/pages/{pageId}* and */pages/{pageId}*.
            Need *include* query param to contain 'content'.

            The content of the page.
          type: string
        contentRevisionId:
          description:
            The page revision used to fill the content.
          $ref: '#/components/schemas/PageRevisionId'
...
    User:
      properties:
        id:
          description: Unique identifier of a user.
          type: string
        reference:
          description: Unique reference if user comes from external source. Use for search only.
          type: string
        first_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        email:
          type: string
        editable_profile:
          description: True if the user can edit the MyAccount information
          type: boolean
          default: false
        permissions:
          description: Map of Permissions.
          type: object
          properties:
            APPLICATION:
              type: array
              items:
                type: string
            USER:
              type: array
              items:
                type: string
        customFields:
          type: object
          description: Values for CustomUserFields
          additionalProperties: true
        config:
          $ref: '#/components/schemas/UserConfig'
        _links:
          $ref: '#/components/schemas/UserLinks'
    UserConfig:
      properties:
        management_url:
          description: The URL of the Gravitee management UI
          type: string
...
    RatingSummary:
      properties:
        average:
          type: number
          format: double
        count:
          type: number
        one_star:
          type: number
          format: double
        two_star:
          type: number
          format: double
        three_star:
          type: number
          format: double
        four_star:
          type: number
          format: double
        five_star:
          type: number
          format: double
...
    MetadataMap:
      description: Map of Map of Object
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: object
...
    FilterApiQuery:
      type: string
      enum:
        - ALL
        - FEATURED
        - MINE
        - STARRED
        - TRENDINGS
...
    Links:
      properties:
        self:
          type: string
        first:
          type: string
        last:
          type: string
        prev:
          type: string
        next:
          type: string
    ApiLinks:
      properties:
        self:
          type: string
        links:
          type: string
        metrics:
          type: string
        pages:
          type: string
        picture:
          type: string
        background:
          type: string
        plans:
          type: string
        ratings:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    APINotFoundError:
      description: API not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PictureSuccess:
      description: A picture
      content:
        'image/*':
          schema:
            type: string
            format: binary
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM